requirements.txt
# AI Customer Research - Python Dependencies

# Core Dependencies
python-dotenv>=1.0.0          # Environment variable management
pyyaml>=6.0.1                 # Configuration file parsing
requests>=2.31.0              # HTTP client for Ollama API

# Data Processing
pandas>=2.1.0                 # Data manipulation and analysis
numpy>=1.24.0                 # Numerical computing

# Text Processing (Optional - uncomment if needed)
# nltk>=3.8.1                 # Natural Language Toolkit
# spacy>=3.7.0                # Industrial-strength NLP
# textblob>=0.17.1            # Simple text processing

# Ollama Integration
ollama>=0.1.0                 # Official Ollama Python client

# Visualization (Optional)
matplotlib>=3.8.0             # Plotting library
seaborn>=0.13.0               # Statistical visualization

# Testing
pytest>=7.4.0                 # Testing framework
pytest-cov>=4.1.0             # Coverage plugin
pytest-asyncio>=0.21.0        # Async test support
pytest-mock>=3.12.0           # Mocking support

# Code Quality
flake8>=6.1.0                 # Linting
black>=23.12.0                # Code formatting
isort>=5.13.0                 # Import sorting
mypy>=1.7.0                   # Type checking

# Logging & Monitoring
loguru>=0.7.2                 # Advanced logging
tqdm>=4.66.0                  # Progress bars

# Async Support
aiohttp>=3.9.0                # Async HTTP client
asyncio>=3.4.3                # Async I/O support

# Development Tools
ipython>=8.18.0               # Enhanced Python shell
jupyter>=1.0.0                # Notebook support (optional)

# Production
gunicorn>=21.2.0              # WSGI server (if building API)
uvicorn>=0.24.0               # ASGI server (if using FastAPI)
